+++ lib/components/search/Search.js
@@ -2,14 +2,14 @@ import React from 'react';
 import AsyncSelect from 'react-select/lib/Async';
 import { withStyles } from '@material-ui/core/styles';
 
-import Placeholder from './search/Placeholder';
-import NoOptionsMessage from './search/NoOptionsMessage';
-import Control from './search/Control';
-import SingleValue from './search/SingleValue';
-import ValueContainer from './search/ValueContainer';
-import DropdownIndicator from './search/DropdownIndicator';
-import Menu from './search/Menu';
-import OptionContainer from './search/OptionContainer';
+import Placeholder from './Placeholder';
+import NoOptionsMessage from './NoOptionsMessage';
+import Control from './Control';
+import SingleValue from './SingleValue';
+import ValueContainer from './ValueContainer';
+import DropdownIndicator from './DropdownIndicator';
+import Menu from './Menu';
+import OptionContainer from './OptionContainer';
 
 const styles = theme => ({
   root: {
