+++ lib/index.js
@@ -25,7 +25,7 @@ export { default as SubHeading } from './components/typography/SubHeading';
 export { default as PageTitle } from './components/typography/PageTitle';
 export { default as Splash } from './components/Splash';
 export { default as HomeBox } from './components/HomeBox';
-export { default as Search } from './components/Search';
+export { default as Search } from './components/search/Search';
 export { default as SearchExamples } from './components/SearchExamples';
 export { default as OtUiThemeProvider } from './components/OtUiThemeProvider';
 export { default as OtTable } from './components/OtTable';
